backend/pyproject.toml
[project]
name = "agentloop-backend"
version = "1.3.0"
description = "AgentLoop — AI Agent 협업 문서 워크스페이스 백엔드"
requires-python = ">=3.13"
dependencies = [
    "fastapi",
    "uvicorn",
    "pyyaml",
    "python-multipart",
]

[project.scripts]
dev = "uvicorn main:app --reload --port 8066"

[dependency-groups]
dev = [
    "httpx>=0.28.1",
]
